register-verification

+++ src/main/java/com/projet/jee/model/DetailService.java
@@ -12,7 +12,7 @@ public class DetailService {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String titre;
     private String description;
@@ -33,8 +33,8 @@ public class DetailService {
     private Set<ServicePhoto> photos;
 
     // === Getters et Setters ===
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getTitre() { return titre; }
     public void setTitre(String titre) { this.titre = titre; }
